Source Code Inspection/src/br/calebe/ticketmachine/core/Troco.java
package br.calebe.ticketmachine.core;

import java.util.Iterator;

/**
 *
 * @author Calebe de Paula Bianchini
 */
public class Troco {

    protected PapelMoeda[] papeisMoeda;

    public Troco(int valor) {
        papeisMoeda = new PapelMoeda[6];
        int[] papelmoeda_valido = {100,50,20,10,5,2};
        int qtd_nota = 0;
        int i = 0;
        while (valor != 0) {
            qtd_nota = valor / papelmoeda_valido[i] ;
            valor = valor % papelmoeda_valido[i];
            papeisMoeda[(papeisMoeda.length - 1) - i] = new PapelMoeda(papelmoeda_valido[i], qtd_nota);
            i++;
        }
    }

    public Iterator<PapelMoeda> getIterator() {
        return new TrocoIterator(this);
    }

    class TrocoIterator implements Iterator<PapelMoeda> {

        protected Troco troco;

        public TrocoIterator(Troco troco) {
            this.troco = troco;
        }

        @Override
        public boolean hasNext() {
            for (int i = 0; i < troco.papeisMoeda.length; i++) {
                if (troco.papeisMoeda[i] != null) {
                    return true;
                }
            }
            return false;
        }

        @Override
        public PapelMoeda next() {
            PapelMoeda ret = null;
            for (int i = 0; i < troco.papeisMoeda.length ; i++) {
                if (troco.papeisMoeda[i] != null) {
                    ret = troco.papeisMoeda[i];
                    troco.papeisMoeda[i] = null;
                }
            }
            return ret;
        }

        @Override
        public void remove() {
            next();
        }
    }
}